地区管理 完善(缺少图片上传) 选区完善 用户管理完善

+++ ui/page_elements/user_manage/__init__.py
@@ -5,7 +5,7 @@ from PyQt5.QtWidgets import (QCheckBox, QHBoxLayout, QHeaderView, QMessageBox,
 from model.user import User
 from ui.page_elements.modal_dialog import ModalDialog
 from ui.page_elements.user_add import UserAdd
-
+from libs.g import g
 from .usermanagerUI import Ui_Dialog
 
 
@@ -46,6 +46,7 @@ class UserManager(ModalDialog):
             item.setTextAlignment(Qt.AlignCenter)
             item.setFlags(Qt.ItemIsEnabled)
             table_widget.setItem(index, 0, item)
+
             item = QTableWidgetItem()
             item.setFont(self.font())
             item.setText(i.nickname)
@@ -67,11 +68,12 @@ class UserManager(ModalDialog):
         self.ui.tableWidget.horizontalHeader().setFont(self.font())
 
     def modify(self):
+        flag=0
         row = self.ui.tableWidget.rowCount()
         for i in range(row):
             username = self.ui.tableWidget.item(i, 0).text()
             nickname = self.ui.tableWidget.item(i, 1).text()
-            ckb = self.ui.tableWidget.item(i, 2)
+            ckb = self.ui.tableWidget.cellWidget(i, 2).children()[1]
             if ckb.checkState() == Qt.Checked:
                 permission = 1
             else:
@@ -80,18 +82,26 @@ class UserManager(ModalDialog):
             if target_user.nickname != nickname:
                 target_user.modify(nickname=nickname)
 
-            if target_user.permission != permission:
-                target_user.modify(permission=permission)
 
-            QMessageBox.information(None, "管理用户", "保存成功")
+            if target_user.permission != permission:
+                if target_user.username!= g.current_user.username :
+                    target_user.modify(permission=permission)
+                else :
+                    flag = 1
+        if  flag==0:
+            QMessageBox.information(None, "管理用户", "保存成功!")
+        else:
+            QMessageBox.information(None, "管理用户", "保存成功,本人权限未修改!")
+        self.refresh()
+
 
     def add_user(self):
+        self.user_add_dialog = UserAdd(self)
         self.user_add_dialog.ui.LineEdit.setText("")
         self.user_add_dialog.ui.LineEdit_2.setText("")
         self.user_add_dialog.ui.LineEdit_3.setText("")
         self.user_add_dialog.ui.LineEdit_4.setText("")
         self.user_add_dialog.show()
-        self.refresh()
 
     def del_user(self):
         select_row = self.ui.tableWidget.currentRow()
@@ -99,7 +109,10 @@ class UserManager(ModalDialog):
             QMessageBox.warning(None, "删除用户失败", "未选定用户!")
         else:
             username = self.ui.tableWidget.item(select_row, 0).text()
-            target_user = User.search(username=username, page_size=-1)["data"][0]
-            target_user.delete()
-            QMessageBox.information(None, "管理用户", "删除成功")
-            self.refresh()
+            if username==g.current_user.username :
+                QMessageBox.warning(None, "删除用户失败", "无法删除自己!")
+            else :
+                target_user = User.search(username=username, page_size=-1)["data"][0]
+                target_user.delete()
+                QMessageBox.information(None, "管理用户", "删除成功")
+                self.refresh()
